@@ -20,7 +20,6 @@ namespace phpseclib3\Crypt\Common\Formats\Keys;
 use phpseclib3\Common\Functions\Strings;
 use phpseclib3\Crypt\AES;
 use phpseclib3\Crypt\Random;
-use phpseclib3\Exception\UnsupportedFormatException;
 
 /**
  * OpenSSH Formatted RSA Key Handler
@@ -67,7 +66,7 @@ abstract class OpenSSH
         // key format is described here:
         // https://cvsweb.openbsd.org/cgi-bin/cvsweb/src/usr.bin/ssh/PROTOCOL.key?annotate=HEAD
 
-        if (strpos($key, 'BEGIN OPENSSH PRIVATE KEY') !== false) {
+        if (str_contains($key, 'BEGIN OPENSSH PRIVATE KEY')) {
             $key = preg_replace('#(?:^-.*?-[\r\n]*$)|\s#ms', '', $key);
             $key = Strings::base64_decode($key);
             $magic = Strings::shift($key, 15);
