@@ -1252,7 +1252,7 @@ abstract class ASN1
             }
             $prefix = substr($content, 0, 2) >= 50 ? '19' : '20';
             $content = $prefix . $content;
-        } elseif (strpos($content, '.') !== false) {
+        } elseif (str_contains($content, '.')) {
             $format .= '.u';
         }
 
@@ -1260,7 +1260,7 @@ abstract class ASN1
             $content = substr($content, 0, -1) . '+0000';
         }
 
-        if (strpos($content, '-') !== false || strpos($content, '+') !== false) {
+        if (str_contains($content, '-') || str_contains($content, '+')) {
             $format .= 'O';
         }
 
