@@ -178,9 +178,7 @@ abstract class TestCase extends PhpseclibTestCase
      */
     public function continuousBufferBatteryCombosWithoutSingleCombos(): array
     {
-        return array_filter($this->continuousBufferBatteryCombos(), function (array $continuousBufferBatteryCombo) {
-            return count($continuousBufferBatteryCombo[2]) > 1;
-        });
+        return array_filter($this->continuousBufferBatteryCombos(), fn (array $continuousBufferBatteryCombo) => count($continuousBufferBatteryCombo[2]) > 1);
     }
 
     /**
