@@ -20,6 +20,6 @@ namespace phpseclib3\Exception;
  *
  * @author  Jim Wigginton <terrafrost@php.net>
  */
-class UnableToConnectException extends \RuntimeException
+class UnableToConnectException extends \RuntimeException implements ExceptionInterface
 {
 }
